+++ tests/testthat/test_tiling.R
@@ -49,6 +49,8 @@ test_that("tiling patterns work as expected", {
         g.ppt("3.3.3.12*30.3.3.12*30", gp = gp3))
     expect_doppelganger("3.3.8*15.3.4.3.8*15", function()
         g.ppt("3.3.8*15.3.4.3.8*15", gp = gp3))
+    expect_doppelganger("3.4.6.3.12*30", function()
+        g.ppt("3.4.6.3.12*30", gp = gp3))
     expect_doppelganger("3.4.8.3.8*15", function()
         g.ppt("3.4.8.3.8*15", gp = gp3))
     expect_doppelganger("4.6*30.4.6*30.4.6*30", function()