+++ NEWS.md
@@ -17,7 +17,7 @@ Breaking Changes
 New Features
 ------------
 
-* `grid.pattern_polygon_tiling()` now supports the following additional polygon tiling `type`'s:
+* `grid.pattern_polygon_tiling()` now supports the following additional polygon tiling `type`'s (#42):
 
   - `"rhombille"` implements a rhombille tiling of rhombi (#37)
   - `"tetrakis_square"` implements a tetrakis square tiling of isosceles right triangles (#38)
@@ -25,6 +25,8 @@ New Features
   - `"3.3.3.12*30.3.3.12*30"` implements a regular (star) polygon tiling made of triangles and twelve-pointed starts.
   - `"3.3.8*15.3.4.3.8*15"` implements a regular (star) polygon tiling of triangles,
     squares, and eight-pointed stars.
+  - `"3.4.6.3.12*30"` implements a regular (star) polygon tiling made of triangles, squares,
+     hexagons, and twelve-pointed stars.
   - `"3.4.8.3.8*15"` implements a regular (star) polygon tiling of triangles,
     squares, octagons, and eight-pointed stars.
   - `"4.6*30.4.6*30.4.6*30"` implements a regular (star) polygon of squares and six-pointed stars.